refactor: api files inside one folder and code indentation
