+++ .gitignore
@@ -1,1 +1,2 @@
 db.sqlite3
+__pycache__/
