+++ src/services/api/manage.py
@@ -1,0 +1,21 @@
+#!/usr/bin/env python
+import os
+import sys
+from pathlib import Path
+sys.path.append(str(Path(__file__).parent.parent.parent))
+
+def main():
+  """Run administrative tasks."""
+  os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'searcher.settings')
+  try:
+    from django.core.management import execute_from_command_line
+  except ImportError as exc:
+    raise ImportError(
+      "Couldn't import Django. Are you sure it's installed and "
+      "available on your PYTHONPATH environment variable? Did you "
+      "forget to activate a virtual environment?"
+    ) from exc
+  execute_from_command_line(sys.argv)
+
+if __name__ == '__main__':
+  main()
