+++ src/services/api/searcher/asgi.py
@@ -1,7 +1,4 @@
-import os
 from django.core.asgi import get_asgi_application
 from django.core.handlers.asgi import ASGIHandler
-
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'django_notifier.settings')
 
 application: ASGIHandler = get_asgi_application()
