+++ src/services/api/searcher/controllers/__init__.py
@@ -1,0 +1,1 @@
+from .get_way import HandleGetWayViewSet
