+++ src/services/api/searcher/controllers/get_way/body_schema.py
@@ -1,0 +1,7 @@
+body_schema = {
+  "type" : "object",
+  "properties" : {
+    "sourceValue" : {"type" : "string"},
+    "targetValue" : {"type" : "string"},
+  },
+}
