+++ src/services/api/searcher/wsgi.py
@@ -1,7 +1,4 @@
-import os
 from django.core.handlers.wsgi import WSGIHandler
 from django.core.wsgi import get_wsgi_application
-
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'django_notifier.settings')
 
 application: WSGIHandler = get_wsgi_application()
